requirements.txt
# Core MCP
mcp

# HTTP RAG backend
fastapi
uvicorn
requests
python-dotenv
pydantic

# Vector store (choose what you actually use)
chromadb

# Utilities
tqdm
